+++ src/components/members/spColDef.tsx
@@ -1,7 +1,16 @@
 import AirtableLinkCell from '@/components/react-table/cells/AirtableLink';
+import HoverCardCell from '@/components/react-table/cells/HoverCardCell';
 import ReactTableRoleCell from '@/components/react-table/cells/Role';
+import TooltipWithLink from '@/components/react-table/cells/TooltipWithLink';
 import { SoloProjectSubmission } from '@/types/SoloProjectTypes';
+import { SiGithub } from '@icons-pack/react-simple-icons';
 import { createColumnHelper } from '@tanstack/table-core';
+import {
+  Globe,
+  MessageCirclePlusIcon,
+  MessageSquareIcon,
+  StepForwardIcon,
+} from 'lucide-react';
 
 const columnHelper =
   createColumnHelper<SoloProjectSubmission>();
@@ -17,6 +26,16 @@ export const spColDef = (baseURL: string) => [
         />
       );
     },
+  }),
+  columnHelper.display({
+    id: 'Solo Project Internal Link',
+    cell: ({ row }) => (
+      <TooltipWithLink
+        link={`/solo-project/${row.original.id}`}
+        tooltip={`Go to Solo Project`}
+        Icon={StepForwardIcon}
+      />
+    ),
   }),
   columnHelper.accessor((row) => row.fields['Timestamp'], {
     id: 'Timestamp',
@@ -43,6 +62,44 @@ export const spColDef = (baseURL: string) => [
       header: 'Status',
     }
   ),
+  columnHelper.display({
+    id: 'Github Repo',
+    cell: ({ row }) => (
+      <TooltipWithLink
+        link={row.original.fields['GitHub Repo URL']}
+        Icon={SiGithub}
+        tooltip="Go to GitHub Repo"
+      />
+    ),
+  }),
+  columnHelper.display({
+    id: 'Deployed App',
+    cell: ({ row }) => (
+      <TooltipWithLink
+        link={row.original.fields['Deployed App URL']}
+        Icon={Globe}
+        tooltip="Go to Deployed App"
+      />
+    ),
+  }),
+  columnHelper.display({
+    id: 'Feedback',
+    cell: ({ row }) => (
+      <HoverCardCell
+        content={row.original.fields['Evaluation Feedback']}
+        Icon={MessageSquareIcon}
+      />
+    ),
+  }),
+  columnHelper.display({
+    id: 'Additional Comments',
+    cell: ({ row }) => (
+      <HoverCardCell
+        content={row.original.fields['Addl. Comments']}
+        Icon={MessageCirclePlusIcon}
+      />
+    ),
+  }),
   columnHelper.display({
     id: 'Tier',
     header: 'Tier',