+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 ## [1.10.0](https://github.com/cherylli/chingu-soloproject-evaluation-app/compare/v1.9.0...v1.10.0) (2025-08-11)
 
-
 ### Features
 
-* **ui:** add `AirtableLinkButton` and `BackButton`, integrate into admin pages and components ([5578893](https://github.com/cherylli/chingu-soloproject-evaluation-app/commit/55788935868916252af3786302341606a01907b5))
-* add Voyage Schedule page which links to each voyage (signups) ([371c5fe](https://github.com/cherylli/chingu-soloproject-evaluation-app/commit/371c5fe3b0278d51de0e781e93dd25e3e6b0f814))
+- **ui:** add `AirtableLinkButton` and `AirtableLinkButton`, integrate into admin pages and components ([5578893](https://github.com/cherylli/chingu-soloproject-evaluation-app/commit/55788935868916252af3786302341606a01907b5))
+- add Voyage Schedule page which links to each voyage (signups) ([371c5fe](https://github.com/cherylli/chingu-soloproject-evaluation-app/commit/371c5fe3b0278d51de0e781e93dd25e3e6b0f814))
 
 ## [1.9.0](https://github.com/cherylli/chingu-soloproject-evaluation-app/compare/v1.8.0...v1.9.0) (2025-08-09)